+++ src/containers/App.js
@@ -113,8 +113,8 @@ class App extends React.Component {
   }
 
   render () {
-    const {activeView, activeModal, actionButton, classes, views, switchView, closeModal, appBarMenuItems,
-      detailViewButtons, settingsTab, appBarButtons, appBarTitle, willPrint, paperSize, cancelPrint, changePaperSize, mapControls} = this.props
+    const {activeView, activeModal, actionButton, mapboxToken, classes, views, switchView, closeModal, appBarMenuItems,
+      detailViewButtons, settingsTab, appBarButtons, appBarTitle, willPrint, paperSize, cancelPrint, changePaperSize} = this.props
     const ViewComponent = getViewComponent(activeView, views)
 
     return (
@@ -129,7 +129,7 @@ class App extends React.Component {
         <div className={classes.inner}>
           <ConnectFilterPane />
           <div className={classes.view}>
-            <ViewContainer component={ViewComponent} mapControls={activeView === 'map' ? mapControls : undefined} />
+            <ViewContainer component={ViewComponent} mapboxToken={mapboxToken} />
           </div>
           {actionButton && <div className={classes.actionButton}>{createElement(actionButton)}</div>}
         </div>